Backend

See merge request s08-webmobile1-sub2/S08P12B304!192

+++ backend/Project/bobs/src/main/java/com/b304/bobs/api/controller/CommunityController.java
@@ -1,5 +1,6 @@
 package com.b304.bobs.api.controller;
 
+import com.b304.bobs.api.request.CommunityDelReq;
 import com.b304.bobs.api.request.CommunityReq;
 import com.b304.bobs.api.response.CommunityRes;
 import com.b304.bobs.api.response.ModifyRes;
@@ -8,12 +9,10 @@ import com.b304.bobs.api.response.PageRes;
 import com.b304.bobs.api.service.CommunityService;
 import com.b304.bobs.util.FileUtils;
 import lombok.RequiredArgsConstructor;
-import org.apache.tika.Tika;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.parameters.P;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
@@ -131,18 +130,27 @@ public class CommunityController {
     }
 
     @PutMapping
-    private ResponseEntity<?> modify(@RequestBody CommunityReq communityReq) {
+    private ResponseEntity<?> modify(CommunityReq communityReq) throws Exception {
         Map<String, Object> map = new HashMap<String, Object>();
+        Long user_id = communityReq.getUser_id();
+        Long community_id =  communityReq.getCommunity_id();
+
+        Long Origin_id = communityService.findOne(community_id).getUser_id();
+        if(!user_id.equals(Origin_id)){
+            map.put("result", false);
+            return ResponseEntity.status(HttpStatus.UNSUPPORTED_MEDIA_TYPE).body(map);
+        }
 
         try {
-            if (!FileUtils.validImgFile(communityReq.getCommunity_img().getInputStream())) {
+            if (!FileUtils.validImgFile(communityReq.getCommunity_img().getInputStream())
+                    && communityReq.getCommunity_img().getSize() != 0) {
                 map.put("result", false);
                 return ResponseEntity.status(HttpStatus.UNSUPPORTED_MEDIA_TYPE).body(map);
             }
 
             ModifyRes modifyRes = communityService.modifyCommunity(communityReq);
             if (modifyRes.getResult()) {
-                map.put("community_id", modifyRes.getId());
+                map.put("data", modifyRes.getObject());
                 map.put("result", true);
                 return ResponseEntity.status(HttpStatus.OK).body(map);
             } else {
@@ -156,12 +164,21 @@ public class CommunityController {
     }
 
     @DeleteMapping()
-    private ResponseEntity<?> delete(@RequestParam(value = "value") Long community_id) {
+    private ResponseEntity<?> delete(@RequestBody CommunityDelReq communityDelReq) throws Exception {
         Map<String, Object> map = new HashMap<String, Object>();
+
+        Long user_id = communityDelReq.getUser_id();
+        Long community_id =  communityDelReq.getCommunity_id();
+
+        Long Origin_id = communityService.findOne(community_id).getUser_id();
+        if(!user_id.equals(Origin_id)){
+            map.put("result", false);
+            return ResponseEntity.status(HttpStatus.UNSUPPORTED_MEDIA_TYPE).body(map);
+        }
+
         try {
             ModifyRes modifyRes = communityService.deleteCommunity(community_id);
             if (modifyRes.getResult()) {
-                map.put("community_id", modifyRes.getId());
                 map.put("result", true);
                 return ResponseEntity.status(HttpStatus.OK).body(map);
             } else {
